requirements.txt
tensorflow-gpu==2.1.0
tensorflow_hub==0.7.0
elasticsearch==7.6.0
pillow==7.0.0
flask==1.1.1
